add observabilityaddon spec nil check.

+++ pkg/controller/placementrule/placementrule_controller.go
@@ -236,7 +236,7 @@ func (r *ReconcilePlacementRule) Reconcile(request reconcile.Request) (reconcile
 		return reconcile.Result{}, nil
 	}
 
-	if !deleteAll && !mco.Spec.ObservabilityAddonSpec.EnableMetrics {
+	if !deleteAll && mco.Spec.ObservabilityAddonSpec != nil && !mco.Spec.ObservabilityAddonSpec.EnableMetrics {
 		reqLogger.Info("EnableMetrics is set to false. Delete Observability addons")
 		deleteAll = true
 	}
